requirements.txt
click==8.1.7
requests==2.32.3
pyyaml==6.0
pytest==7.1.2
PyJWT==2.10.1
shodan==1.30.0
python-nmap==0.7.1
beautifulsoup4==4.12.3
cryptography==42.0.8
ldap3==2.9